Extract InlineNote component for shared styles
